refactor(app): HelpService の HTML ヘルプウィンドウ管理を HelpHtmlWindowController へ分離 (#208)
WebView2 ベースのヘルプ HTML ウィンドウのライフサイクル管理とナビゲーション
イベントハンドリングを HelpHtmlWindowController へ移動し、HelpService は
「どのヘルプを表示するか」のオーケストレーションに専念させる（559行 → 299行）。

URI 解決の純粋関数群は現状維持（本 Issue の分割対象外）。
