+++ PhotoGeoExplorer/Services/HelpService.cs
@@ -4,14 +4,9 @@ using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.IO;
 using System.Threading.Tasks;
-using Microsoft.UI.Windowing;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
-using Microsoft.Web.WebView2.Core;
 using Microsoft.Windows.Globalization;
-using Windows.Graphics;
-using Windows.System;
-using WinRT.Interop;
 
 namespace PhotoGeoExplorer.Services;
 
@@ -25,12 +20,8 @@ internal sealed class HelpService : IHelpService
     private readonly Action<bool> _setShowQuickStartOnStartup;
     private readonly Func<Task> _saveQuickStartPreferenceAsync;
     private readonly bool _settingsFileExistsAtStartup;
-    private readonly Action<string, InfoBarSeverity> _showNotification;
+    private readonly HelpHtmlWindowController _helpHtmlWindowController;
     private bool _disposed;
-    private Window? _helpHtmlWindow;
-    private WebView2? _helpHtmlWebView;
-    private Uri? _helpHtmlLocalFallbackUri;
-    private Uri? _helpHtmlExternalBaseUri;
 
     public HelpService(
         IDialogService dialogService,
@@ -49,7 +40,8 @@ internal sealed class HelpService : IHelpService
         _setShowQuickStartOnStartup = setShowQuickStartOnStartup ?? throw new ArgumentNullException(nameof(setShowQuickStartOnStartup));
         _saveQuickStartPreferenceAsync = saveQuickStartPreferenceAsync ?? throw new ArgumentNullException(nameof(saveQuickStartPreferenceAsync));
         _settingsFileExistsAtStartup = settingsFileExistsAtStartup;
-        _showNotification = showNotification ?? throw new ArgumentNullException(nameof(showNotification));
+        ArgumentNullException.ThrowIfNull(showNotification);
+        _helpHtmlWindowController = new HelpHtmlWindowController(_getExternalContentBaseUrl, showNotification);
     }
 
     public Task ShowGettingStartedAsync()
@@ -82,38 +74,15 @@ internal sealed class HelpService : IHelpService
             return;
         }
 
-        _helpHtmlLocalFallbackUri = localFallbackUri;
-        _helpHtmlExternalBaseUri = externalUri is null
+        var externalBaseUri = externalUri is null
             ? null
             : GetExternalContentBaseUri(_getExternalContentBaseUrl());
 
-        if (_helpHtmlWindow is not null)
-        {
-            if (_helpHtmlWebView is not null)
-            {
-                _helpHtmlWebView.Source = uri;
-            }
-
-            _helpHtmlWindow.Activate();
-            return;
-        }
-
-        var webView = CreateHelpHtmlWebView(uri);
-        _helpHtmlWebView = webView;
-
-        var container = new Grid();
-        container.Children.Add(webView);
-
-        var window = new Window
-        {
-            Title = LocalizationService.GetString("Dialog.Help.Html.Title"),
-            Content = container
-        };
-        window.Closed += (_, _) => CleanupHelpHtmlWindow();
-
-        _helpHtmlWindow = window;
-        window.Activate();
-        TryResizeHelpWindow(window, 980, 720);
+        _helpHtmlWindowController.ShowOrNavigate(
+            uri,
+            localFallbackUri,
+            externalBaseUri,
+            LocalizationService.GetString("Dialog.Help.Html.Title"));
     }
 
     public async Task ShowAboutAsync()
@@ -138,7 +107,7 @@ internal sealed class HelpService : IHelpService
 
     public void CloseHelpWindow()
     {
-        CloseHelpHtmlWindow();
+        _helpHtmlWindowController.Close();
     }
 
     public void Dispose()
@@ -149,7 +118,7 @@ internal sealed class HelpService : IHelpService
         }
 
         _disposed = true;
-        CloseHelpWindow();
+        _helpHtmlWindowController.Dispose();
         GC.SuppressFinalize(this);
     }
 
@@ -327,233 +296,5 @@ internal sealed class HelpService : IHelpService
         await _dialogService.ShowMessageDialogAsync(
             LocalizationService.GetString("Dialog.Help.HtmlMissing.Title"),
             LocalizationService.GetString("Dialog.Help.HtmlMissing.Detail")).ConfigureAwait(true);
-    }
-
-    private WebView2 CreateHelpHtmlWebView(Uri uri)
-    {
-        var webView = new WebView2
-        {
-            Source = uri,
-            HorizontalAlignment = HorizontalAlignment.Stretch,
-            VerticalAlignment = VerticalAlignment.Stretch
-        };
-
-        webView.NavigationStarting += OnHelpWebViewNavigationStarting;
-        webView.NavigationCompleted += OnHelpWebViewNavigationCompleted;
-        webView.CoreWebView2Initialized += OnHelpWebViewInitialized;
-        return webView;
-    }
-
-    private void CleanupHelpHtmlWindow()
-    {
-        CloseHelpHtmlWebView();
-        _helpHtmlWindow = null;
-    }
-
-    private void CloseHelpHtmlWindow()
-    {
-        if (_helpHtmlWindow is null)
-        {
-            CleanupHelpHtmlWindow();
-            return;
-        }
-
-        try
-        {
-            _helpHtmlWindow.Close();
-        }
-        catch (Exception ex) when (ex is InvalidOperationException
-            or System.Runtime.InteropServices.COMException
-            or UnauthorizedAccessException)
-        {
-            AppLog.Error("Failed to close help window.", ex);
-            CleanupHelpHtmlWindow();
-        }
-    }
-
-    private void CloseHelpHtmlWebView()
-    {
-        if (_helpHtmlWebView is null)
-        {
-            return;
-        }
-
-        try
-        {
-            _helpHtmlWebView.NavigationStarting -= OnHelpWebViewNavigationStarting;
-            _helpHtmlWebView.NavigationCompleted -= OnHelpWebViewNavigationCompleted;
-            _helpHtmlWebView.CoreWebView2Initialized -= OnHelpWebViewInitialized;
-            if (_helpHtmlWebView.CoreWebView2 is not null)
-            {
-                _helpHtmlWebView.CoreWebView2.NewWindowRequested -= OnHelpWebViewNewWindowRequested;
-            }
-
-            _helpHtmlWebView.Close();
-        }
-        catch (Exception ex) when (ex is InvalidOperationException
-            or System.Runtime.InteropServices.COMException)
-        {
-            AppLog.Error("Failed to close help WebView2.", ex);
-        }
-        finally
-        {
-            _helpHtmlWebView = null;
-            _helpHtmlLocalFallbackUri = null;
-            _helpHtmlExternalBaseUri = null;
-        }
-    }
-
-    private void OnHelpWebViewInitialized(WebView2 sender, CoreWebView2InitializedEventArgs args)
-    {
-        if (args.Exception is not null)
-        {
-            AppLog.Error("Help WebView2 initialization failed.", args.Exception);
-            return;
-        }
-
-        sender.CoreWebView2.NewWindowRequested += OnHelpWebViewNewWindowRequested;
-    }
-
-    private void OnHelpWebViewNewWindowRequested(object? sender, CoreWebView2NewWindowRequestedEventArgs args)
-    {
-        if (TryGetExternalUri(args.Uri, out var uri) && uri is not null)
-        {
-            if (ShouldOpenOutsideHelpWindow(uri))
-            {
-                args.Handled = true;
-                _ = OpenExternalUriAsync(ResolveExternalUri(uri));
-                return;
-            }
-
-            args.Handled = true;
-            if (_helpHtmlWebView is not null)
-            {
-                _helpHtmlWebView.Source = uri;
-            }
-        }
-    }
-
-    private void OnHelpWebViewNavigationStarting(WebView2 sender, CoreWebView2NavigationStartingEventArgs args)
-    {
-        if (TryGetExternalUri(args.Uri, out var uri) && uri is not null)
-        {
-            if (ShouldOpenOutsideHelpWindow(uri))
-            {
-                args.Cancel = true;
-                _ = OpenExternalUriAsync(ResolveExternalUri(uri));
-            }
-        }
-    }
-
-    private void OnHelpWebViewNavigationCompleted(WebView2 sender, CoreWebView2NavigationCompletedEventArgs args)
-    {
-        if (args.IsSuccess || _helpHtmlExternalBaseUri is null || _helpHtmlLocalFallbackUri is null)
-        {
-            return;
-        }
-
-        if (sender.Source is null || AreSameUri(sender.Source, _helpHtmlLocalFallbackUri))
-        {
-            return;
-        }
-
-        AppLog.Info("External help page load failed. Falling back to local help content.");
-        _helpHtmlExternalBaseUri = null;
-        sender.Source = _helpHtmlLocalFallbackUri;
-    }
-
-    private bool ShouldOpenOutsideHelpWindow(Uri uri)
-    {
-        if (_helpHtmlExternalBaseUri is null)
-        {
-            return true;
-        }
-
-        return !AreSameOrigin(uri, _helpHtmlExternalBaseUri);
-    }
-
-    private static bool AreSameOrigin(Uri left, Uri right)
-    {
-        return string.Equals(left.Scheme, right.Scheme, StringComparison.OrdinalIgnoreCase)
-            && string.Equals(left.Host, right.Host, StringComparison.OrdinalIgnoreCase)
-            && left.Port == right.Port;
-    }
-
-    private static bool AreSameUri(Uri left, Uri right)
-    {
-        return string.Equals(left.AbsoluteUri, right.AbsoluteUri, StringComparison.OrdinalIgnoreCase);
-    }
-
-    private Uri ResolveExternalUri(Uri uri)
-    {
-        if (uri.AbsolutePath.Contains("privacy-policy", StringComparison.OrdinalIgnoreCase))
-        {
-            return GetExternalPrivacyPolicyUri(_getExternalContentBaseUrl()) ?? uri;
-        }
-
-        return uri;
-    }
-
-    private static bool TryGetExternalUri(string? uriString, out Uri? uri)
-    {
-        uri = null;
-        if (string.IsNullOrWhiteSpace(uriString))
-        {
-            return false;
-        }
-
-        if (!Uri.TryCreate(uriString, UriKind.Absolute, out var parsed))
-        {
-            return false;
-        }
-
-        if (string.Equals(parsed.Scheme, Uri.UriSchemeHttp, StringComparison.OrdinalIgnoreCase)
-            || string.Equals(parsed.Scheme, Uri.UriSchemeHttps, StringComparison.OrdinalIgnoreCase))
-        {
-            uri = parsed;
-            return true;
-        }
-
-        return false;
-    }
-
-    private async Task OpenExternalUriAsync(Uri uri)
-    {
-        try
-        {
-            await Launcher.LaunchUriAsync(uri);
-        }
-        catch (Exception ex) when (ex is InvalidOperationException
-            or UnauthorizedAccessException
-            or System.Runtime.InteropServices.COMException
-            or ArgumentException)
-        {
-            AppLog.Error("Failed to launch help link.", ex);
-            _showNotification(
-                LocalizationService.GetString("Message.LaunchBrowserFailed"),
-                InfoBarSeverity.Error);
-        }
-    }
-
-    private static void TryResizeHelpWindow(Window window, int width, int height)
-    {
-        try
-        {
-            var appWindow = GetAppWindow(window);
-            appWindow.Resize(new SizeInt32(width, height));
-        }
-        catch (Exception ex) when (ex is ArgumentException
-            or InvalidOperationException
-            or System.Runtime.InteropServices.COMException)
-        {
-            AppLog.Error("Failed to resize help window.", ex);
-        }
-    }
-
-    private static AppWindow GetAppWindow(Window window)
-    {
-        var hwnd = WindowNative.GetWindowHandle(window);
-        var windowId = Microsoft.UI.Win32Interop.GetWindowIdFromWindow(hwnd);
-        return AppWindow.GetFromWindowId(windowId);
     }
 }
